requirements.txt
openai>=1.0.0
click>=8.0.0
rich>=13.0.0
pyyaml>=6.0
aiohttp>=3.8.0
requests>=2.31.0
python-dotenv>=0.19.0
prometheus-client>=0.16.0
structlog>=22.1.0
pytest>=7.3.1
pytest-asyncio>=0.20.3
black>=22.3.0
mypy>=1.0.0
pylint>=2.17.0